feat(game): implement crime point accumulation and wanted status logic
- Apply the "Wanted" buff automatically when a character's crime points reach the defined threshold (50).
- Implement `SCCrimeChangedPacket` to synchronize crime and infamy point updates with the client.
- Add support for updating crime points for offline characters via direct database persistence.
- Refactor crime point values from `int` to `short` across managers, quests, and doodad functions to match protocol expectations.
- Update `SCCharacterStatePacket` to correctly report jury points.
- Refactor and expand `BuffConstants` to include "Wanted" status buff and updated some others.

+++ AAEmu.Game/Core/Managers/UnitManagers/DoodadManager.cs
@@ -1103,7 +1103,7 @@ public class DoodadManager(IObjectIdManager objectIdManager, IDoodadIdManager do
                         {
                             Id = reader.GetUInt32("id"),
                             SkillId = reader.GetUInt32("skill_id"),
-                            CrimeValue = reader.GetInt32("crime_value"),
+                            CrimeValue = reader.GetInt16("crime_value"),
                             CrimeKindId = reader.GetUInt32("crime_kind_id")
                         };
                         _funcTemplates["DoodadFuncEvidenceItemLoot"].Add(func.Id, func);
