+++ AAEmu.Game/Models/Game/DoodadObj/Funcs/DoodadFuncEvidenceItemLoot.cs
@@ -7,7 +7,7 @@ public class DoodadFuncEvidenceItemLoot : DoodadFuncTemplate
 {
     // doodad_funcs
     public uint SkillId { get; set; }
-    public int CrimeValue { get; set; }
+    public short CrimeValue { get; set; }
     public uint CrimeKindId { get; set; }
 
     public override void Use(BaseUnit caster, Doodad owner, uint skillId, int nextPhase = 0)
